# src/backend/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Car Share API
  description: This is the API for the Car Share project.
  version: 0.4.0 # TODO: Update version number
  contact:
    name: Car Share, backend team
    email: 'contoux1u@etu.univ-lorraine.fr'
  license:
    name: GNU General Public License v3.0
    url: 'https://www.gnu.org/licenses/gpl-3.0.en.html'
servers:
  - url: 'https://carshare.api.antoinectx.fr'
    description: Antoine's server
  - url: 'https://carshare.apis.furwaz.fr'
    description: Paul's server
...
paths:
...
  /users/password-reset:
    post:
      summary: Ask for password reset
      description: Send an email to reset the password
      operationId: askForPasswordReset
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: john@example.com
      responses:
        '200':
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: If the provided email address matches an account, a password reset email has been sent.
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email is required.
        '429':
          $ref: '#/components/responses/tooManyRequests'
        '500':
          $ref: '#/components/responses/internalError'
    patch:
      summary: Password reset
      description: Change the password
      operationId: resetPassword
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      security:
        - bearerAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: Password123$
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password changed
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password is required.
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /users/email-verification:
    post:
      summary: Ask for email verification
      description: Send an email to verify the email address
      operationId: askForEmailVerification
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: A verification link has been sent to your email address.
        '401':
          $ref: '#/components/responses/noToken'
        '429':
          $ref: '#/components/responses/tooManyRequests'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    patch:
      summary: Email verification
      description: Verify the email address
      operationId: emailVerification
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email address verified
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /users:
    get:
      summary: Search users (admin)
      description: "Search in all users. Only available to admins.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: searchUsers
      tags:
        - Users
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/query'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrivateUser'
                  query:
                    type: string
                    required: false
                    example: friends
                  limit:
                    type: integer
                    example: 10
                  offset:
                    type: integer
                    example: 0
                  next:
                    type: integer
                    nullable: true
                    example: 10
                  prev:
                    type: integer
                    nullable: true
                    example: 0
                  total:
                    type: integer
                    example: 12
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /users/search:
    get:
      summary: Search users (public)
      description: Search in all users. Available to everyone.
      operationId: searchUsersPublic
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/query'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: A list of minimal information about users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MinimalUser'
                  query:
                    type: string
                    required: false
                    example: John
                  limit:
                    type: integer
                    example: 10
                  offset:
                    type: integer
                    example: 0
                  next:
                    type: integer
                    nullable: true
                    example: 10
                  prev:
                    type: integer
                    nullable: true
                    example: 0
                  total:
                    type: integer
                    example: 12
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /users/:id:
    delete:
      summary: Delete a user (admin)
      description: "Delete a user. Only available to admins.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: deleteUser
      tags:
        - Users
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the user to delete
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User deleted
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '404':
          $ref: '#/components/responses/userNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    patch:
      summary: Update a user (admin)
      description: "Update a user. Only available to admins.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: updateUser
      tags:
        - Users
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      requestBody:
        description: The user to update
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  required: false
                  example: john@example.com
                firstName:
                  type: string
                  required: false
                  example: John
                lastName:
                  type: string
                  required: false
                  example: Doe
                phone:
                  type: string
                  required: false
                  example: 0123456789
                gender:
                  type: integer
                  required: false
                  example: 1
                hasCar:
                  type: boolean
                  required: false
                  example: true
                level:
                  type: integer
                  required: false
                  example: 1
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    properties:
                      message:
                        type: string
                        example: User updated
                      user:
                        $ref: '#/components/schemas/PrivateUser'
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '404':
          $ref: '#/components/responses/userNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  parameters:
    Accept-Language:
      name: Accept-Language
      in: header
      description: The language to use in the response message (if any)
      required: false
      schema:
        type: string
        default: en
        enum:
          - en
          - fr
    limit:
      name: limit
      in: query
      description: The number of items to return
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 50
    offset:
      name: offset
      in: query
      description: The number of items to skip before starting to collect the result set
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    query:
      name: query
      in: query
      description: The query to search for
      required: false
      schema:
        type: string
        default: ''
  responses:
...
    invalidId:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Id is invalid.
    noToken:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: No token provided.
    invalidToken:
      description: Invalid/Expired token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid token provided.
    insufficientPrivileges:
      description: Insufficient privileges
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: You do not have sufficient rights to perform this action.
...
    userNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: User not found.
...
    tooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: An email has already been sent to this address less than 10 minutes ago. Please wait before trying again.
    internalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: An internal error occurred. If the problem persists, please contact an administrator.
...
  schemas:
...
    MinimalUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: john@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
...
    PrivateUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: john@example.com
        emailVerifiedOn:
          type: string
          format: date-time
          nullable: true
          default: null
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phone:
          type: string
          example: "+330612345678"
        avatar:
          type: string
          nullable: true
          default: null
        gender:
          type: integer
          default: -1
        hasCar:
          type: boolean
          default: false
        mailNotif:
          type: boolean
          default: false
        level:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        locale:
          type: string
          nullable: true
          default: null
          example: fr
        timezone:
          type: string
          nullable: true
          default: null
          example: Europe/Paris
...